# openapi.yaml
openapi: 3.1.0
info:
  title: AV Suite Backend Foundation
  description: Rehabilitation exercises management backend platform
  version: 0.1.0
paths:
...
  /api/v1/patients:
    get:
      tags:
      - Patients
      - Patients
      summary: List Patients
      description: "Endpoint ka purpose: Clinic ke sare patients ko list karna pagination\
        \ ke saath\nYeh endpoint patient listing handle karta hai jismein:\n1. Authorization\
        \ check (physio ya admin)\n2. Clinic-specific patients fetch karte hain\n\
        3. Pagination apply karte hain\n4. Metadata ke saath results return karte\
        \ hain\n\nHTTP Method: GET\nURL: /api/v1/patients\nStatus Code: 200 OK\n\n\
        Query Parameters (PaginationParams):\n- page: Page number (1-based, default:\
        \ 1)\n- page_size: Results per page (default: 10, max: 100)\n\nRequest Headers:\n\
        Authorization: Bearer {jwt_token}\n\nResponse (ResponseEnvelope[List[PatientRead]]):\n\
        {\n    \"status\": \"success\",\n    \"data\": [\n        {\n            \"\
        id\": \"uuid\",\n            \"clinic_id\": \"uuid\",\n            \"first_name\"\
        : \"John\",\n            \"last_name\": \"Doe\",\n            \"date_of_birth\"\
        : \"1990-01-01\",\n            \"phone\": \"9876543210\",\n            \"\
        created_at\": \"2024-01-01T00:00:00Z\",\n            \"updated_at\": \"2024-01-01T00:00:00Z\"\
        \n        },\n        ...\n    ],\n    \"meta\": {\n        \"total\": 45,\n\
        \        \"page\": 1,\n        \"page_size\": 10\n    }\n}\n\nError Responses:\n\
        - 401 UNAUTHORIZED: Missing or invalid JWT token\n- 403 FORBIDDEN: User role\
        \ (patient, nurse) is not authorized\n- 400 BAD REQUEST: Invalid pagination\
        \ parameters\n- 500 INTERNAL_SERVER_ERROR: Database error\n\nSecurity:\n-\
        \ JWT token required: All requests authenticated hone chahiye\n- Clinic isolation:\
        \ Sirf user ke clinic ke patients visible\n- Role check: Only admin/physio\
        \ access kar sakte hain\n\nPagination:\n- Server-side pagination: Large datasets\
        \ efficiently handle\n- Total count: Frontend pagination UI ke liye\n- Page\
        \ size limit: DOS attack prevention\n\nUsage Example:\ncurl -X GET \"http://localhost:8000/api/v1/patients?page=1&page_size=10\"\
        \ \\\n  -H \"Authorization: Bearer eyJhbGc...\""
      operationId: list_patients_api_v1_patients_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: Page number
          default: 1
          title: Page
        description: Page number
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Items per page
          default: 20
          title: Page Size
        description: Items per page
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEnvelope_List_PatientRead__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Patients
      - Patients
      summary: Create Patient
      description: "Endpoint ka purpose: New patient create karna clinic mein\nYeh\
        \ endpoint patient registration handle karta hai.\n\nHTTP Method: POST\nURL:\
        \ /api/v1/patients\nStatus Code: 201 CREATED (new patient created)\n\nRequest\
        \ Body (PatientCreate):\n{\n    \"first_name\": \"John\",              # Required\n\
        \    \"last_name\": \"Doe\",                # Required\n    \"date_of_birth\"\
        : \"1990-01-01\",     # Optional\n    \"phone\": \"9876543210\"          \
        \    # Optional\n}\n\nRequest Headers:\nAuthorization: Bearer {jwt_token}\n\
        Content-Type: application/json\n\nResponse (ResponseEnvelope[PatientRead]):\n\
        {\n    \"status\": \"success\",\n    \"data\": {\n        \"id\": \"new-uuid\"\
        ,\n        \"clinic_id\": \"clinic-uuid\",\n        \"first_name\": \"John\"\
        ,\n        \"last_name\": \"Doe\",\n        \"date_of_birth\": \"1990-01-01\"\
        ,\n        \"phone\": \"9876543210\",\n        \"created_at\": \"2024-01-01T00:00:00Z\"\
        ,\n        \"updated_at\": \"2024-01-01T00:00:00Z\"\n    }\n}\n\nError Responses:\n\
        - 401 UNAUTHORIZED: Missing or invalid JWT token\n- 403 FORBIDDEN: User role\
        \ not authorized for patient creation\n- 400 BAD REQUEST: Invalid request\
        \ body or validation failure\n- 500 INTERNAL_SERVER_ERROR: Database error\n\
        \nSecurity:\n- JWT token required: All requests authenticated\n- Clinic isolation:\
        \ Patient automatically current clinic se associated\n- Role check: Only admin/physio\
        \ create kar sakte hain\n\nBusiness Logic:\n1. Authorization check - physio\
        \ ya admin\n2. Request data validation (Pydantic schema)\n3. Patient database\
        \ record create\n4. Created patient return\n\nUsage Example:\ncurl -X POST\
        \ http://localhost:8000/api/v1/patients \\\n  -H \"Authorization: Bearer eyJhbGc...\"\
        \ \\\n  -H \"Content-Type: application/json\" \\\n  -d '{\n    \"first_name\"\
        : \"John\",\n    \"last_name\": \"Doe\",\n    \"date_of_birth\": \"1990-01-01\"\
        ,\n    \"phone\": \"9876543210\"\n  }'"
      operationId: create_patient_api_v1_patients_post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEnvelope_PatientRead_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{id}:
    get:
      tags:
      - Patients
      - Patients
      summary: Get Patient
      description: "Endpoint ka purpose: Specific patient ke details fetch karna\n\
        Yeh endpoint patient details retrieve karta hai by ID.\n\nHTTP Method: GET\n\
        URL: /api/v1/patients/{id}\nStatus Code: 200 OK (found)\n\nPath Parameters:\n\
        - id (str): Patient UUID\n\nRequest Headers:\nAuthorization: Bearer {jwt_token}\n\
        \nResponse (ResponseEnvelope[PatientRead]):\n{\n    \"status\": \"success\"\
        ,\n    \"data\": {\n        \"id\": \"patient-uuid\",\n        \"clinic_id\"\
        : \"clinic-uuid\",\n        \"first_name\": \"John\",\n        \"last_name\"\
        : \"Doe\",\n        \"date_of_birth\": \"1990-01-01\",\n        \"phone\"\
        : \"9876543210\",\n        \"created_at\": \"2024-01-01T00:00:00Z\",\n   \
        \     \"updated_at\": \"2024-01-01T00:00:00Z\"\n    }\n}\n\nError Responses:\n\
        - 401 UNAUTHORIZED: Missing or invalid JWT token\n- 403 FORBIDDEN: User role\
        \ not authorized\n- 404 NOT_FOUND: Patient not found or not in user's clinic\n\
        - 400 BAD REQUEST: Invalid patient ID format\n- 500 INTERNAL_SERVER_ERROR:\
        \ Database error\n\nSecurity:\n- JWT token required: All requests authenticated\n\
        - Clinic isolation: Patient sirf user ke clinic mein\n- Role check: Only admin/physio\
        \ access kar sakte hain\n- Cross-clinic prevention: Can't access other clinic's\
        \ patients\n\nBusiness Logic:\n1. Authorization check - physio ya admin\n\
        2. Patient fetch by ID aur clinic check\n3. Patient found = return data, Not\
        \ found = 404\n\nUsage Example:\ncurl -X GET http://localhost:8000/api/v1/patients/12345678-1234-5678-1234-567812345678\
        \ \\\n  -H \"Authorization: Bearer eyJhbGc...\""
      operationId: get_patient_api_v1_patients__id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEnvelope_PatientRead_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Patients
      - Patients
      summary: Update Patient
      operationId: update_patient_api_v1_patients__id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEnvelope_PatientRead_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Patients
      - Patients
      summary: Delete Patient
      operationId: delete_patient_api_v1_patients__id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEnvelope_NoneType_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MetaError:
      properties:
        error:
          $ref: '#/components/schemas/MetaErrorDetails'
      type: object
      required:
      - error
      title: MetaError
    MetaErrorDetails:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        field:
          anyOf:
          - type: string
          - type: 'null'
          title: Field
      type: object
      required:
      - code
      - message
      title: MetaErrorDetails
    MetaPagination:
      properties:
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
      type: object
      required:
      - total
      - page
      - page_size
      title: MetaPagination
...
    PatientCreate:
      properties:
        first_name:
          type: string
          title: First Name
        last_name:
          type: string
          title: Last Name
        date_of_birth:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Date Of Birth
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        user_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: User Id
      type: object
      required:
      - first_name
      - last_name
      title: PatientCreate
...
    PatientRead:
      properties:
        first_name:
          type: string
          title: First Name
        last_name:
          type: string
          title: Last Name
        date_of_birth:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Date Of Birth
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        id:
          type: string
          format: uuid
          title: Id
        clinic_id:
          type: string
          format: uuid
          title: Clinic Id
        user_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: User Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - first_name
      - last_name
      - id
      - clinic_id
      - user_id
      - created_at
      - updated_at
      title: PatientRead
    PatientUpdate:
      properties:
        first_name:
          anyOf:
          - type: string
          - type: 'null'
          title: First Name
        last_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Name
        date_of_birth:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Date Of Birth
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        user_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: User Id
      type: object
      title: PatientUpdate
...
    ResponseEnvelope_List_PatientRead__:
      properties:
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/PatientRead'
            type: array
          - type: 'null'
          title: Data
        meta:
          anyOf:
          - $ref: '#/components/schemas/MetaPagination'
          - $ref: '#/components/schemas/MetaError'
          - type: object
          - type: 'null'
          title: Meta
      type: object
      title: ResponseEnvelope[List[PatientRead]]
...
    ResponseEnvelope_NoneType_:
      properties:
        data:
          type: 'null'
          title: Data
        meta:
          anyOf:
          - $ref: '#/components/schemas/MetaPagination'
          - $ref: '#/components/schemas/MetaError'
          - type: object
          - type: 'null'
          title: Meta
      type: object
      title: ResponseEnvelope[NoneType]
    ResponseEnvelope_PatientRead_:
      properties:
        data:
          anyOf:
          - $ref: '#/components/schemas/PatientRead'
          - type: 'null'
        meta:
          anyOf:
          - $ref: '#/components/schemas/MetaPagination'
          - $ref: '#/components/schemas/MetaError'
          - type: object
          - type: 'null'
          title: Meta
      type: object
      title: ResponseEnvelope[PatientRead]
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login